# packages/contracts/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: AEP BFF API
  version: 1.0.0
servers:
- url: /api/v1
paths:
...
  /projects/{projectName}/issues:
    get:
      operationId: list-issues
      summary: List/search GitHub issues on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      - description: Comma-separated GitHub labels to filter by
        in: query
        name: labels
        schema:
          type: string
      - description: Keyword search over issue title/body, ranked by distinct-term
          overlap (title weighted double), capped at 25. Recall-biased — used to
          surface related issues before filing a new one.
        in: query
        name: q
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/IssueInfo'
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
    post:
      operationId: create-issue
      summary: Create a GitHub issue on a project's repo
      tags:
      - Issues
      security:
      - userJWT: []
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResult'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
...
  /projects/{projectName}/tasks:
    get:
      operationId: list-tasks
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: Which Tasks to return (default open)
        explode: false
        in: query
        name: state
        schema:
          description: Which Tasks to return (default open)
          enum:
          - open
          - closed
          - all
          type: string
      - description: Filter to the Tasks of one spec/build version tag (e.g. v3). The tag is resolved
          to a milestone number through the platform's run rows and the filter is milestone
          MEMBERSHIP — never a title match against GitHub. Empty returns every version.
        explode: false
        in: query
        name: tag
        schema:
          description: Filter to the Tasks of one spec/build version tag (e.g. v3), by milestone
            membership. Empty returns every version.
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TaskView'
                type: array
                nullable: true
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: List a project's issues (live from GitHub), optionally scoped to one spec/build version
        tag by milestone membership. Every row carries the label-derived `executorClass` the console
        sections on — agent work, a dispatch gate, or a bare human ledger issue. The aep:validation
        issue is excluded (it is a phase of the run, surfaced with the run's verdict on the deployment
        surface); get-task still serves it by number. Ledger issues have no label to query by, so they
        are only discoverable through milestone membership and appear only on a `tag`-scoped read.
      tags:
      - Tasks
  /projects/{projectName}/tasks/{issueNumber}:
    get:
      operationId: get-task
      parameters:
      - description: Project name (DNS-label slug)
        in: path
        name: projectName
        required: true
        schema:
          description: Project name (DNS-label slug)
          type: string
      - description: GitHub issue number of the Task
        in: path
        name: issueNumber
        required: true
        schema:
          description: GitHub issue number of the Task
          format: int64
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error
      security:
      - userJWT: []
      summary: Get one Task with its full Execution history
      tags:
      - Tasks
...
components:
  schemas:
...
    Error:
      description: Flat error envelope returned by every non-2xx response.
      type: object
      additionalProperties: false
      properties:
        code:
          description: Machine-readable error code (stable slug).
          type: string
          example: validation_failed
        message:
          description: Human-readable error message.
          type: string
          example: request body is invalid
        details:
          description: Field-level details for validation errors.
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
      required:
      - code
      - message
    ErrorDetail:
      description: One field-level validation error.
      type: object
      additionalProperties: false
      properties:
        field:
          description: Location of the offending field (e.g. body.name).
          type: string
        message:
          description: What is wrong with the field.
          type: string
      required:
      - field
      - message
    ExecutionView:
      additionalProperties: false
      properties:
        createdAt:
          format: date-time
          type: string
        endedAt:
          format: date-time
          type: string
          nullable: true
          x-go-type: '*time.Time'
        id:
          type: string
        kind:
          type: string
        reason:
          type: string
        runName:
          type: string
        startedAt:
          format: date-time
          type: string
          nullable: true
          x-go-type: '*time.Time'
        status:
          type: string
      required:
      - id
      - kind
      - status
      - createdAt
      type: object
...
    CreateIssueRequest:
      additionalProperties: false
      description: Issue to file on the project's repo. dedupeKey makes creation
        idempotent per open issue (label-encoded), for concurrent alert handlers.
        Adoption is the DEFAULT — an issue filed here is agent work unless the
        caller opts out with adopt=false.
      properties:
        title:
          type: string
        body:
          type: string
        labels:
          items:
            type: string
          type: array
          nullable: true
        dedupeKey:
          type: string
        adopt:
          # Keep an explicit pointer under the global prefer-skip-optional-pointer:
          # absent must read as the default (true), not as the zero value (false),
          # or every caller that omits the field would silently stop adopting.
          #
          # The default is stated in prose rather than as `default: true` on
          # purpose: openapi-typescript renders a default-bearing property as
          # REQUIRED, which would force every TS caller to pass the very field
          # whose whole point is that omitting it does the right thing.
          x-go-type-skip-optional-pointer: false
          description: Whether to hand the issue to the coding agent. OMITTED
            MEANS TRUE — the issue joins the deployed version's milestone as
            agent work and a run picks it up. false files a ledger issue instead
            — recorded against the version, never worked until somebody adopts
            it.
          type: boolean
        componentName:
          description: The DESIGN component this issue is about, unprefixed (e.g.
            `service1`, not `myproject-service1`). Optional, and read only when
            adopting — it provisions the component's OpenChoreo Component CR
            up-front, so a name the design does not carry fails THIS call instead
            of surfacing later inside a cycle. Omit it and adoption proceeds
            without the check.
          type: string
      required:
      - title
      - body
      type: object
    IssueResult:
      additionalProperties: false
      description: Issue metadata after create. deduped=true means an open issue
        with the same dedupeKey already existed — number/url refer to it.
        adopted answers the only other question a caller has — will anything
        work this issue?
      properties:
        number:
          format: int64
          type: integer
        url:
          type: string
        nodeId:
          type: string
        deduped:
          type: boolean
        adopted:
          description: Whether the issue was adopted — filed into a version's
            milestone as agent work, with a run started or woken over it. False
            when the caller passed adopt=false, when creation deduped onto an
            existing issue (the run that created it owns its dispatch), or when
            adoption could not proceed; adoptionError says which.
          type: boolean
        adoptionError:
          description: Why adoption did not happen, when it was asked for and
            did not. The issue still exists — it is a ledger entry, and adding
            the `aep:codingagent` label hands it to the coding agent later.
          type: string
      required:
      - number
      - url
      - nodeId
      type: object
    IssueInfo:
      additionalProperties: false
      description: 'One issue from list/search. Field names are CAPITALIZED on
        the wire (historical shape the deployed aep-mcp-server parses — do not
        "fix" without a coordinated MCP-server release).'
      properties:
        Number:
          format: int64
          type: integer
        Title:
          type: string
        Body:
          type: string
        URL:
          type: string
        State:
          type: string
        Labels:
          items:
            type: string
          type: array
          nullable: true
      required:
      - Number
      - Title
      - Body
      - URL
      - State
      - Labels
      type: object
...
    Lineage:
      additionalProperties: false
      properties:
        designTag:
          type: string
        specTag:
          type: string
      type: object
...
    TaskDetail:
      additionalProperties: false
      properties:
        blockedBy:
          description: Names of the dependencies this task is waiting on; present when derivedStatus is
            on_hold.
          items:
            type: string
          type: array
          nullable: true
        attention:
          items:
            type: string
          type: array
          nullable: true
        body:
          type: string
        component:
          type: string
        dependsOn:
          items:
            type: string
          type: array
          nullable: true
        derivedStatus:
          type: string
        executionHistory:
          items:
            $ref: '#/components/schemas/ExecutionView'
          type: array
          nullable: true
        executions:
          additionalProperties:
            $ref: '#/components/schemas/ExecutionView'
          type: object
        executorClass:
          description: >-
            Label-derived kind of the issue, and the only classification the
            platform makes of one: `coding` for agent work (the `aep` label),
            `provision` for a dispatch gate (`aep:provision`), `validation` for
            the run's validation issue, `ledger` for a bare human issue that
            joined the milestone carrying none of them. Nothing here is parsed
            out of the body — issue bodies are prose the platform writes for
            the agent and never reads back.
          enum:
          - coding
          - provision
          - validation
          - ledger
          type: string
        hold:
          type: boolean
        issueNumber:
          format: int64
          type: integer
        issueUrl:
          type: string
        lineage:
          $ref: '#/components/schemas/Lineage'
        operation:
          type: string
        origin:
          type: string
        rationale:
          type: string
        title:
          type: string
      required:
      - executionHistory
      - issueNumber
      - title
      - issueUrl
      - executorClass
      - dependsOn
      - lineage
      - derivedStatus
      - hold
      - attention
      - executions
      type: object
...
    TaskView:
      additionalProperties: false
      properties:
        attention:
          items:
            type: string
          type: array
          nullable: true
        blockedBy:
          description: Names of the dependencies this task is waiting on; present when derivedStatus is
            on_hold.
          items:
            type: string
          type: array
        body:
          type: string
        component:
          type: string
        dependsOn:
          items:
            type: string
          type: array
          nullable: true
        derivedStatus:
          type: string
        executions:
          additionalProperties:
            $ref: '#/components/schemas/ExecutionView'
          type: object
        executorClass:
          description: >-
            Label-derived kind of the issue, and the only classification the
            platform makes of one: `coding` for agent work (the `aep` label),
            `provision` for a dispatch gate (`aep:provision`), `validation` for
            the run's validation issue, `ledger` for a bare human issue that
            joined the milestone carrying none of them. Nothing here is parsed
            out of the body — issue bodies are prose the platform writes for
            the agent and never reads back.
          enum:
          - coding
          - provision
          - validation
          - ledger
          type: string
        hold:
          type: boolean
        issueNumber:
          format: int64
          type: integer
        issueUrl:
          type: string
        lineage:
          $ref: '#/components/schemas/Lineage'
        operation:
          type: string
        origin:
          type: string
        rationale:
          type: string
        title:
          type: string
        usage:
          $ref: "#/components/schemas/Usage"
          description: Aggregate actual usage across the task's executions (#245); absent for tasks that predate usage capture.
      required:
      - issueNumber
      - title
      - issueUrl
      - executorClass
      - dependsOn
      - lineage
      - derivedStatus
      - hold
      - attention
      - executions
      type: object
...
    Usage:
      additionalProperties: false
      description: Actual token usage for one unit of agent work or an aggregate (#245, #291). Tokens + model are the persisted truth; costUsd is stamped at capture time from the model rates then in force (ADR-0011 as amended) and null when the row predates stamping or its model had no rate. Aggregates sum stamps and never reprice.
      properties:
        cacheCreationTokens:
          format: int64
          type: integer
        cacheReadTokens:
          format: int64
          type: integer
        costUsd:
          description: Write-time-stamped USD (sum, for aggregates); null when no stamp exists.
          format: double
          type: number
          nullable: true
          x-go-type: '*float64'
        inputTokens:
          format: int64
          type: integer
        model:
          description: Model id the work ran on; "" on mixed-model aggregates.
          type: string
        outputTokens:
          format: int64
          type: integer
      required:
        - inputTokens
        - outputTokens
        - cacheReadTokens
        - cacheCreationTokens
        - model
        - costUsd
      type: object
...
  securitySchemes:
    userJWT:
      bearerFormat: JWT
      description: Thunder end-user JWT (RS256, JWKS-verified)
      scheme: bearer
      type: http